+++ library/log/timber/src/main/java/com/freeletics/coredux/log/timber/TimberLogSink.kt
@@ -1,16 +1,16 @@
-package com.freeletics.coredux.log.android
+package com.freeletics.coredux.log.timber
 
-import android.util.Log
 import com.freeletics.coredux.log.common.LoggerLogSink
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.GlobalScope
+import timber.log.Timber
 
 /**
- * Provides [com.freeletics.coredux.LogSink] implementation, that uses Android [Log] class to send log output.
+ * Provides [com.freeletics.coredux.LogSink] implementation, that uses [Timber] class to send log output.
  *
  * @param scope to receive incoming log messages, default is [GlobalScope]
  */
-class AndroidLogSink(
+class TimberLogSink(
     scope: CoroutineScope = GlobalScope
 ) : LoggerLogSink(scope) {
     override fun debug(
@@ -18,7 +18,8 @@ class AndroidLogSink(
         message: String,
         throwable: Throwable?
     ) {
-        Log.d(tag, message, throwable)
+        Timber.tag(tag)
+        Timber.d(throwable, message)
     }
 
     override fun info(
@@ -26,7 +27,8 @@ class AndroidLogSink(
         message: String,
         throwable: Throwable?
     ) {
-        Log.i(tag, message, throwable)
+        Timber.tag(tag)
+        Timber.i(throwable, message)
     }
 
     override fun warning(
@@ -34,6 +36,7 @@ class AndroidLogSink(
         message: String,
         throwable: Throwable?
     ) {
-        Log.w(tag, message, throwable)
+        Timber.tag(tag)
+        Timber.w(throwable, message)
     }
 }