+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+- Add common [LogSink] implementation
+- Add [Timber](https://github.com/JakeWharton/timber) [LogSink] implementation
+
 ## [1.0.0] - 2019-03-26
 - Add initial core library implementation
 - Add initial android log artifact implementation